QuantTrader_Premium/requirements.txt
# Core Libraries
pandas
numpy
matplotlib
seaborn

# API and Environment
python-dotenv
requests

# RL and Modeling
torch
ray[rllib]
gymnasium

# Jupyter Notebooks
notebook

# Testing
pytest

# Optional (for logging)
pyyaml
